+++ models/ai/vars.go
@@ -10,7 +10,7 @@ const (
 )
 
 // Weights for stability map
-var StabilityMap = [8][8]int{
+var StabilityMap = [8][8]int16{
 	{4, -3, 2, 2, 2, 2, -3, 4},
 	{-3, -4, -1, -1, -1, -1, -4, -3},
 	{2, -1, 1, 0, 0, 1, -1, 2},
